# server/src/routes/swagger.yaml
openapi: "3.0.0"
info:
  title: ExamHub API
  description: 師大資工考古題系統 API
  version: 0.1.0
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /exam:
    get:
      tags:
      - Exam
      summary: 拿取所有考古題
      parameters:
      - in: query
        name: courseId
        schema:
          type: number
        description: 拿取特定的課程的考古題
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    docId:
                      type: number
                - $ref: '#/components/schemas/Exam'
    post:
      tags:
      - Exam
      summary: 新增考古題
      requestBody:
        description: 欲建立的考古題
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamCreate'
      responses:
        200:
          description: 成功
  /exam/{docId}:
    get:
      tags:
      - Exam
      summary: 拿取特定的考古題
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: number
          description: 欲拿取的考古題 docId
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
    patch:
      tags:
      - Exam
      summary: 修改特定的考古題
      parameters:
        - in: path
          name: docId
          required: true
          schema:
            type: number
          description: 欲修改的考古題 docId
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamCreate'
      responses:
        200:
          description: 成功
...
components:
  schemas:
...
    ExamCreate:
      properties:
        courseId:
          type: number
        year:
          type: number
        semester:
          type: number
        title:
          type: string
        description:
          type: string
    Exam:
      allOf:
        - $ref: '#/components/schemas/ExamCreate'
        - type: object
          properties:
            userId:
              type: number
            createTime:
              type: string
              format: date-time
            lastUpdateTime:
              type: string
              format: date-time
            upvote:
              type: integer
            downvote:
              type: integer
            folderPath:
              type: string
...